SDK update generated by liblab

+++ src/main/java/47/one0/http/HubsClient.java
@@ -6,8 +6,8 @@ import 47.one0.exceptions.ApiException;
 
 public interface HubsClient {
 
-   ResponseWithHeaders<String> getGlobalHubs(Float count, Float onlyTransient) throws ApiException;
-   ResponseWithHeaders<String> getLibraryHubs(Float sectionId, Float count, Float onlyTransient) throws ApiException;
+   ResponseWithHeaders<String> getGlobalHubs(Float count, Integer onlyTransient) throws ApiException;
+   ResponseWithHeaders<String> getLibraryHubs(Float sectionId, Float count, Integer onlyTransient) throws ApiException;
 
 
 }
